+++ packages/instant-apps-components/src/components/instant-apps-language-translator/support/resources.ts
@@ -1,0 +1,5 @@
+export const CKEditorStyles = `
+  .instant-apps-language-translator-item__nested-input .ck.ck-reset.ck-editor.ck-rounded-corners {
+    margin-left: 30px !important;
+  }
+`;